requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-treebeard==4.3
djangorestframework==3.9.4
filelock==3.0.10
idna==2.8
inflection==0.3.1
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.13.3
lxml==4.3.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tox==3.10.0
traitlets==4.3.2
urllib3==1.24.3
virtualenv==16.5.0
wcwidth==0.1.7
